feat: 优化服务的标签编辑逻辑 & 抽离出标签编辑组件
